+++ crates/miden-testing/src/kernel_tests/tx/test_note.rs
@@ -15,6 +15,7 @@ use miden_lib::{
 use miden_objects::{
     Digest, EMPTY_WORD, ONE, WORD_SIZE,
     account::{AccountBuilder, AccountId},
+    assembly::diagnostics::miette,
     note::{
         Note, NoteAssets, NoteExecutionHint, NoteExecutionMode, NoteInputs, NoteMetadata,
         NoteRecipient, NoteScript, NoteTag, NoteType,
@@ -428,7 +429,7 @@ fn test_note_setup() {
 }
 
 #[test]
-fn test_note_script_and_note_args() {
+fn test_note_script_and_note_args() -> miette::Result<()> {
     let note_args = [
         [Felt::new(91), Felt::new(91), Felt::new(91), Felt::new(91)],
         [Felt::new(92), Felt::new(92), Felt::new(92), Felt::new(92)],
@@ -468,11 +469,13 @@ fn test_note_script_and_note_args() {
     );
 
     tx_context.set_tx_args(tx_args);
-    let process = tx_context.execute_code(code).unwrap();
+    let process = tx_context.execute_code(code)?;
 
     assert_eq!(process.stack.get_word(0), note_args[0]);
 
     assert_eq!(process.stack.get_word(1), note_args[1]);
+
+    Ok(())
 }
 
 fn note_setup_stack_assertions(process: &Process, inputs: &TransactionContext) {
@@ -645,7 +648,7 @@ fn test_get_current_script_root() {
 }
 
 #[test]
-fn test_build_note_metadata() {
+fn test_build_note_metadata() -> miette::Result<()> {
     let tx_context = TransactionContextBuilder::with_standard_account(ONE)
         .with_mock_notes_preserved()
         .build();
@@ -691,7 +694,7 @@ fn test_build_note_metadata() {
             tag = test_metadata.tag(),
         );
 
-        let process = tx_context.execute_code(&code).unwrap();
+        let process = tx_context.execute_code(&code)?;
 
         let metadata_word = [
             process.stack.get(3),
@@ -702,6 +705,8 @@ fn test_build_note_metadata() {
 
         assert_eq!(Word::from(test_metadata), metadata_word, "failed in iteration {iteration}");
     }
+
+    Ok(())
 }
 
 /// This serves as a test that setting a custom timestamp on mock chain blocks works.